// src/app/model/protuit.ts
/*export class Produit {
    id: number | undefined;
    code: string | undefined;
    designation: string | undefined;
    prix: number | undefined;
  
    constructor() {
      this.id = undefined;
      this.code = undefined;
      this.designation = undefined;
      this.prix = undefined;
    }
  }*/
  /*import { categorie } from './Categorie';

export class Produit {
  id: number | undefined;
  code: string | undefined;
  designation: string | undefined;
  prix: number | undefined;
  categorie: Categorie | undefined;
}*/
import { Categorie } from './Categorie';

export class Produit {
  id: number | undefined;
  code: string | undefined;
  designation: string | undefined;
  prix: number | undefined;
  categorie: Categorie | undefined;
}